Move embedding generation to SDK as well
