+++ libs/@local/hash-isomorphic-utils/src/ai-inference-types.ts
@@ -169,18 +169,6 @@ export type ExternalInputWebsocketRequestMessage = {
 export type InferenceWebsocketServerMessage =
   ExternalInputWebsocketRequestMessage;
 
-export type CreateEmbeddingsParams = {
-  input: string[];
-};
-
-export type CreateEmbeddingsReturn = {
-  embeddings: number[][];
-  usage: {
-    prompt_tokens: number;
-    total_tokens: number;
-  };
-};
-
 export type GetResultsFromCancelledInferenceRequestQuery = QueryDefinition<
   InferEntitiesReturn,
   never,
